Refactor server setup: add logging and error handling middleware, implement health check routes, and restore database connection test functionality.
