+++ server/src/middleware/errorHandler.ts
@@ -1,0 +1,11 @@
+import { Request, Response, NextFunction } from "express";
+
+export function errorHandler(
+  err: any,
+  req: Request,
+  res: Response,
+  next: NextFunction
+) {
+  console.error(err.stack);
+  res.status(500).send("Something went wrong!");
+}
