+++ server/src/middleware/logging.ts
@@ -1,0 +1,18 @@
+import { Request, Response, NextFunction, Application } from "express";
+
+export function setupLogging(app: Application, serverStartTime: Date) {
+  app.use((req: Request, res: Response, next: NextFunction) => {
+    const timestamp = new Date().toISOString();
+
+    console.log(`[${timestamp}] ${req.method} ${req.url} - IP: ${req.ip}`);
+
+    const start = Date.now();
+    res.on("finish", () => {
+      const duration = Date.now() - start;
+      console.log(
+        `[${timestamp}] ${req.method} ${req.url} - ${res.statusCode} - ${duration}ms`
+      );
+    });
+    next();
+  });
+}
